[REF] web_tour: move getConsumeEventType function
Move exported function that is used only once in the right file.
