+++ addons/web_tour/static/src/tour_service/tour_compilers.js
@@ -8,7 +8,7 @@ import { tourState } from "./tour_state";
 import * as hoot from "@odoo/hoot-dom";
 import { session } from "@web/session";
 import { setupEventActions } from "@web/../lib/hoot-dom/helpers/events";
-import { callWithUnloadCheck, getConsumeEventType, getScrollParent } from "./tour_utils";
+import { callWithUnloadCheck, getScrollParent } from "./tour_utils";
 import { utils } from "@web/core/ui/ui_service";
 import { TourHelpers } from "./tour_helpers";
 
@@ -28,6 +28,70 @@ import { TourHelpers } from "./tour_helpers";
  * @property {showPointerDuration} number
  * @property {*} pointer - used for controlling the pointer of the tour
  */
+
+/**
+ * @param {HTMLElement} [element]
+ * @param {RunCommand} [runCommand]
+ * @returns {string}
+ */
+function getConsumeEventType(element, runCommand) {
+    if (!element) {
+        return "click";
+    }
+    const { classList, tagName, type } = element;
+    const tag = tagName.toLowerCase();
+
+    // Many2one
+    if (classList.contains("o_field_many2one")) {
+        return "autocompleteselect";
+    }
+
+    // Inputs and textareas
+    if (
+        tag === "textarea" ||
+        (tag === "input" &&
+            (!type ||
+                [
+                    "email",
+                    "number",
+                    "password",
+                    "search",
+                    "tel",
+                    "text",
+                    "url",
+                    "date",
+                    "range",
+                ].includes(type)))
+    ) {
+        if (
+            utils.isSmall() &&
+            element.closest(".o_field_widget")?.matches(".o_field_many2one, .o_field_many2many")
+        ) {
+            return "click";
+        }
+        return "input";
+    }
+
+    // Drag & drop run command
+    if (typeof runCommand === "string" && /^drag_and_drop/.test(runCommand)) {
+        // this is a heuristic: the element has to be dragged and dropped but it
+        // doesn't have class 'ui-draggable-handle', so we check if it has an
+        // ui-sortable parent, and if so, we conclude that its event type is 'sort'
+        if (element.closest(".ui-sortable")) {
+            return "sort";
+        }
+        if (
+            (/^drag_and_drop_native/.test(runCommand) && classList.contains("o_draggable")) ||
+            element.closest(".o_draggable") ||
+            element.draggable
+        ) {
+            return "pointerdown";
+        }
+    }
+
+    // Default: click
+    return "click";
+}
 
 /**
  * @param {string} selector - any valid Hoot selector
